be/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.2
dj-database-url==1.3.0
Django==4.1.5
django-auto-prefetch==0.0.6
django-cors-headers==3.11.0
django-soft-delete==0.9.18
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
pycryptodome==3.17
PyJWT==2.3.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.5.0
tzdata==2022.7
whitenoise==6.4.0
